+++ src/main/java/com/example/social_network_fpt_be/repository/VideoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface VideoRepository extends JpaRepository<Video, Integer>{
+public interface VideoRepository extends JpaRepository<Video, Long>{
 
 }
